Refactor Registry to behave like an Adapter
This commit implements an adapter pattern for `Registry`, and
subsequently move the current in-memory registry from `Registry` to
`Adapter::Memory`.

By refactoring to this pattern it allows us to plug in different backend
to store the events (e.g: `redis`)
