+++ lib/rocketman/event.rb
@@ -4,11 +4,11 @@ module Rocketman
       @event = event
       @payload = payload
       @test = payload.fetch(:test, false)
-      Rocketman::Registry.instance.register_event(event)
+      Rocketman::Registry.register_event(event)
     end
 
     def notify_consumers
-      consumers = Rocketman::Registry.instance.get_consumers_for(@event)
+      consumers = Rocketman::Registry.get_consumers_for(@event)
 
       consumers.each do |consumer, action|
         consumer.instance_exec(@payload, &action)
