+++ spec/spec_helper.rb
@@ -1,7 +1,7 @@
 require "bundler/setup"
 require "pry"
 require "rocketman"
-require "support/test_pool.rb"
+require "support/test_pool"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -13,10 +13,6 @@ RSpec.configure do |config|
   # Swap it out to a stub implementation because Thread environment is hard to test
   Rocketman.send(:remove_const, "Pool")
   Rocketman::Pool = Rocketman::TestPool
-
-  config.before(:each) do
-    Singleton.__init__(Rocketman::Registry)
-  end
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
